pyproject.toml
[project]
name = "anime-llm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=0.5.23",
    "huggingface-hub>=0.26.5",
    "numpy<2.0.0",
    "pandas>=2.2.3",
    "python-dotenv>=1.0.1",
    "ruff>=0.8.2",
    "sentence-transformers>=3.3.1",
    "torch<2.3.0",
]


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
